Support watch-only logins
Watch-only username and password are stored in the config directory to
make it analagous to mnemonic and pin logins. A new option --watch-only
is added (equivalently --auth watch-only) which reads the username and
password from the config dir and performs a watch-only login.

There was the potential for some confusion in the interface because
setmnemonic updated the mnemonic stored in the local config dir, whereas
setwatchonly maps to the server call which updates the credentials on
the server. setmnemonic has been replaced by the set command which is
also used to locally set the watch-only credentials.

green-cli set username xxx
green-cli set password xxx
green-cli set mnemonic "xxx yyy zzz...."
green-cli set mnemonic -f filename

The implementation has also been refactored and there is a common
mechanism so that if the credentials are not present in the config dir
the user is prompted on the command line.