enable gdk shared create_tx implementation via GA_SHARE_TX_IMPL env var
The value of the environment variable is set into the config for
gdk_init, it should be set to "true"/"t"/"1" to enable this feature.

Enablement is via environment variable rather than a command line
option as it will become the default and only behaviour in due course,
and we do not want users to have to make changes to their scripts to
add/remove any flag.