add a getnextevent command
The existing event commands are not suitable for scripted use in a repl
loop: 'getlatestevent' is lossy, e.g. two notifications may arrive close
together and the first could be missed, while 'listen' is intended for
monitoring. In contrast, 'getnextevent' just pops off the events in FIFO
order allowing the caller to filter and wait using criteria that do not
need to be part of the call arguments/options.