Explicitly check on enable telegram
Make an explicit client side check when enabling telegram 2fa that
something else is enabled rather than relying on the server removing
telegram from the config. Having the server remove the config entry is
more confusing than an explicit check (why is telegram not available? is
it because of my config, or the environment?).